only allow backup dotfile if it doesnt already exist

+++ dotfiles.rb
@@ -40,7 +40,7 @@ def dot_file_replace
       # move file to ~/.mks-dotfiles/backups/timestamp
       backup_dir = "#{DOT_DIR}/backups/#{Time.now.strftime("%Y%m%d")}"
       FileUtils.mkdir_p(backup_dir) unless File.exists?(backup_dir)
-      FileUtils.mv(dot_file_location, backup_dir)
+      FileUtils.mv(dot_file_location, backup_dir) unless File.exists?("#{backup_dir}/#{dot_file}")
       FileUtils.ln_s("#{DOT_DIR}/home/#{dot_file}", dot_file_location)
     end
   end
